requirements.txt
asgiref==3.7.2
autopep8==1.7.0
Django
django-cors-headers==4.3.1
djangorestframework==3.14.0
django-rest-knox==4.2.0
drf-yasg==1.21.7
gunicorn
Pillow==10.1.0
pycodestyle==2.9.1
python-decouple==3.8
pytz==2022.2.1
sqlparse==0.4.3
toml==0.10.2
tzdata==2022.6